requirements.txt
discord.py
python-dotenv
google-api-python-client
requests
